refactor(module-4): consolidate messages/messaging dirs, decompose chat composer and bubble, fix as any casts, extract UnallocatedMembersModal — Part of #58
