--- src/features/messages/index.ts
@@ -1,4 +1,0 @@
-export { useMessages } from './hooks/use-messages';
-export { useSendMessage } from './hooks/use-send-message';
-export { useUnreadCount } from './hooks/use-unread-count';
-export type { Message } from './types/message.model';
