--- src/features/messages/types/index.ts
@@ -1,7 +1,0 @@
-export type {
-  MessageDTO,
-  MessagesResponseDTO,
-  UnreadCountResponseDTO,
-} from './message.dto';
-
-export type { Message } from './message.model';
