+++ src/features/messaging/components/chat-composer.tsx
@@ -1,13 +1,9 @@
 import Feather from '@expo/vector-icons/Feather';
 import * as ImagePicker from 'expo-image-picker';
-import type React from 'react';
 import { forwardRef, useImperativeHandle, useMemo, useRef, useState } from 'react';
 import {
   Alert,
-  Image,
-  Modal,
   Pressable,
-  ScrollView,
   TextInput,
   View,
 } from 'react-native';
@@ -31,14 +27,14 @@ import type {
   PickedChatPhoto,
   RecentWorkout,
 } from '../types/messaging.model';
+import { getWorkoutDeepLinkLabel } from '../utils/messaging-format';
 import {
-  getWorkoutDeepLinkLabel,
-  getWorkoutDisplayName,
-} from '../utils/messaging-format';
-import { MessagingChip } from './messaging-chip';
+  ComposerModifierStrip,
+  ComposerActionChips,
+  ComposerPickerModals,
+} from './composer-picker-modal';
 
 type ComposerModal = 'quick' | 'mention' | 'link' | null;
-
 /** Imperative handle exposed to parent via ref */
 export interface ChatComposerHandle {
   focusInput: () => void;
@@ -309,117 +305,41 @@ export const ChatComposer = forwardRef<ChatComposerHandle, ChatComposerProps>(fu
         </View>
       ) : null}
 
-      {deepLink || photo || visibility === 'captains_only' || isAnnouncement || isImportant ? (
-        <ScrollView horizontal showsHorizontalScrollIndicator={false} className="mb-2">
-          <View className="flex-row gap-2 pr-4">
-            {isAnnouncement ? (
-              <MessagingChip
-                label="Announcement"
-                icon="radio"
-                selected
-                tone="amber"
-                onPress={() => setIsAnnouncement(false)}
-              />
-            ) : null}
-            {visibility === 'captains_only' ? (
-              <MessagingChip
-                label={currentRole === 'player' ? 'DM to Captain' : 'Captains Only'}
-                icon="shield"
-                selected
-                onPress={() => setVisibility('all')}
-              />
-            ) : null}
-            {isImportant ? (
-              <MessagingChip
-                label="Important"
-                icon="alert-triangle"
-                selected
-                tone="danger"
-                onPress={() => setIsImportant(false)}
-              />
-            ) : null}
-            {deepLink ? (
-              <MessagingChip
-                label="Link attached"
-                icon="link"
-                selected
-                tone="brand"
-                onPress={() => setDeepLink(null)}
-              />
-            ) : null}
-            {photo ? (
-              <MessagingChip
-                label="Photo attached"
-                icon="image"
-                selected
-                tone="brand"
-                onPress={() => setPhoto(null)}
-              />
-            ) : null}
-          </View>
-        </ScrollView>
-      ) : null}
+      <ComposerModifierStrip
+        currentRole={currentRole}
+        isAnnouncement={isAnnouncement}
+        isImportant={isImportant}
+        isCaptainsOnly={visibility === 'captains_only'}
+        deepLink={deepLink}
+        photo={photo}
+        onClearAnnouncement={() => setIsAnnouncement(false)}
+        onClearCaptainsOnly={() => setVisibility('all')}
+        onClearImportant={() => setIsImportant(false)}
+        onClearDeepLink={() => setDeepLink(null)}
+        onClearPhoto={() => setPhoto(null)}
+      />
 
-      {photo ? (
-        <View className="mb-2 flex-row items-center gap-3">
-          <Image
-            source={{ uri: photo.uri }}
-            style={{ width: 54, height: 54, borderRadius: 10 }}
-            resizeMode="cover"
-          />
-          <AppText className="flex-1 text-xs text-muted" numberOfLines={1}>
-            {photo.name}
-          </AppText>
-        </View>
-      ) : null}
-
-      <ScrollView horizontal showsHorizontalScrollIndicator={false} className="mb-2">
-        <View className="flex-row gap-2 pr-4">
-          <MessagingChip
-            label={visibility === 'captains_only' ? 'Captains' : 'All'}
-            icon={visibility === 'captains_only' ? 'shield' : 'globe'}
-            selected={visibility === 'captains_only'}
-            onPress={() =>
-              setVisibility((current) =>
-                current === 'captains_only' ? 'all' : 'captains_only',
-              )
-            }
-          />
-          {isLeader ? (
-            <>
-              <MessagingChip
-                label="Announcement"
-                icon="radio"
-                selected={isAnnouncement}
-                tone="amber"
-                onPress={() => setIsAnnouncement((current) => !current)}
-              />
-              <MessagingChip
-                label="Important"
-                icon="alert-triangle"
-                selected={isImportant}
-                tone="danger"
-                onPress={() => setIsImportant((current) => !current)}
-              />
-            </>
-          ) : null}
-          <MessagingChip label="Quick" icon="message-square" onPress={() => setModal('quick')} />
-          <MessagingChip label="Mention" icon="at-sign" onPress={() => setModal('mention')} />
-          <MessagingChip label="Link" icon="link" onPress={() => setModal('link')} />
-          {showPhotoAction ? (
-            <MessagingChip label="Photo" icon="image" onPress={handlePickPhoto} />
-          ) : null}
-          {isCaptainRole && teamId ? (
-            <MessagingChip
-              label={motivating ? 'Writing...' : 'Send Motivation'}
-              icon="zap"
-              disabled={motivating}
-              onPress={handleMotivate}
-              tone="amber"
-            />
-          ) : null}
-        </View>
-      </ScrollView>
+      <ComposerActionChips
+        visibility={visibility}
+        isAnnouncement={isAnnouncement}
+        isImportant={isImportant}
+        isLeader={isLeader}
+        isCaptainRole={isCaptainRole}
+        motivating={motivating}
+        showPhotoAction={showPhotoAction}
+        hasTeam={!!teamId}
+        currentRole={currentRole}
+        onToggleVisibility={() =>
+          setVisibility((current) => (current === 'captains_only' ? 'all' : 'captains_only'))
+        }
+        onToggleAnnouncement={() => setIsAnnouncement((current) => !current)}
+        onToggleImportant={() => setIsImportant((current) => !current)}
+        onOpenQuick={() => setModal('quick')}
+        onOpenMention={() => setModal('mention')}
+        onOpenLink={() => setModal('link')}
+        onPickPhoto={handlePickPhoto}
+        onMotivate={handleMotivate}
+      />
 
       <View className="flex-row items-end gap-2">
         <TextInput
@@ -450,173 +370,20 @@ export const ChatComposer = forwardRef<ChatComposerHandle, ChatComposerProps>(fu
         </Button>
       </View>
 
-      <PickerModal
-        visible={modal === 'quick'}
-        title="Quick Messages"
+      <ComposerPickerModals
+        modal={modal}
+        cannedLoading={cannedQuery.isLoading}
+        cannedMessages={cannedQuery.data ?? []}
+        membersLoading={membersQuery.isLoading}
+        visibleMembers={visibleMembers}
+        workoutsLoading={workoutsQuery.isLoading}
+        workouts={workoutsQuery.data ?? []}
         onClose={() => setModal(null)}
-      >
-        <MessagePickerContent
-          isLoading={cannedQuery.isLoading}
-          emptyLabel="No quick messages."
-          items={cannedQuery.data ?? []}
-          renderItem={(message) => (
-            <PickerRow
-              key={message.id}
-              title={message.title}
-              subtitle={message.content}
-              onPress={() => insertCannedMessage(message)}
-            />
-          )}
-        />
-      </PickerModal>
-
-      <PickerModal
-        visible={modal === 'mention'}
-        title="Mention Member"
-        onClose={() => setModal(null)}
-      >
-        <MessagePickerContent
-          isLoading={membersQuery.isLoading}
-          emptyLabel="No members found."
-          items={visibleMembers}
-          renderItem={(member) => (
-            <PickerRow
-              key={member.userId}
-              title={`@${member.username}`}
-              subtitle={member.roles.join(', ') || 'player'}
-              onPress={() => insertMention(member)}
-            />
-          )}
-        />
-      </PickerModal>
-
-      <PickerModal
-        visible={modal === 'link'}
-        title="Attach Link"
-        onClose={() => setModal(null)}
-      >
-        <View className="gap-2">
-          <PickerRow
-            title="Challenges"
-            subtitle="League challenges"
-            onPress={() => attachStaticLink('challenges')}
-          />
-          <PickerRow
-            title="Leaderboard"
-            subtitle="Current standings"
-            onPress={() => attachStaticLink('leaderboard')}
-          />
-          <PickerRow
-            title="Activities"
-            subtitle="Activity configuration"
-            onPress={() => attachStaticLink('activities')}
-          />
-          <AppText className="px-1 pt-2 text-xs font-semibold text-muted">
-            Recent Activities
-          </AppText>
-          <MessagePickerContent
-            isLoading={workoutsQuery.isLoading}
-            emptyLabel="No recent workouts."
-            items={workoutsQuery.data ?? []}
-            renderItem={(workout) => (
-              <PickerRow
-                key={workout.id}
-                title={getWorkoutDisplayName(workout)}
-                subtitle={`${workout.status.replace(/_/g, ' ')} | ${new Date(workout.date).toLocaleDateString()}`}
-                onPress={() => attachWorkout(workout)}
-              />
-            )}
-          />
-        </View>
-      </PickerModal>
+        onSelectCanned={insertCannedMessage}
+        onSelectMember={insertMention}
+        onAttachStaticLink={attachStaticLink}
+        onAttachWorkout={attachWorkout}
+      />
     </View>
   );
 });
-
-function PickerModal({
-  visible,
-  title,
-  children,
-  onClose,
-}: {
-  visible: boolean;
-  title: string;
-  children: React.ReactNode;
-  onClose: () => void;
-}) {
-  return (
-    <Modal visible={visible} transparent animationType="fade" onRequestClose={onClose}>
-      <View className="flex-1 justify-end" style={{ backgroundColor: 'rgba(0,0,0,0.28)' }}>
-        <View
-          className="max-h-[70%] rounded-t-3xl bg-card p-4"
-          style={{ borderTopWidth: 1, borderColor: mflColors.border }}
-        >
-          <View className="mb-3 flex-row items-center gap-3">
-            <AppText className="flex-1 text-base font-semibold text-foreground">
-              {title}
-            </AppText>
-            <Pressable onPress={onClose} hitSlop={10}>
-              <Feather name="x" size={22} color={mflColors.textMuted} />
-            </Pressable>
-          </View>
-          <ScrollView showsVerticalScrollIndicator={false}>{children}</ScrollView>
-        </View>
-      </View>
-    </Modal>
-  );
-}
-
-function MessagePickerContent<T>({
-  isLoading,
-  emptyLabel,
-  items,
-  renderItem,
-}: {
-  isLoading: boolean;
-  emptyLabel: string;
-  items: T[];
-  renderItem: (item: T) => React.ReactNode;
-}) {
-  if (isLoading) {
-    return (
-      <View className="items-center py-8">
-        <Spinner size="sm" />
-      </View>
-    );
-  }
-
-  if (items.length === 0) {
-    return (
-      <View className="items-center py-8">
-        <AppText className="text-sm text-muted">{emptyLabel}</AppText>
-      </View>
-    );
-  }
-
-  return <View className="gap-2">{items.map(renderItem)}</View>;
-}
-
-function PickerRow({
-  title,
-  subtitle,
-  onPress,
-}: {
-  title: string;
-  subtitle: string;
-  onPress: () => void;
-}) {
-  return (
-    <Pressable
-      onPress={onPress}
-      className="rounded-xl border px-3 py-3"
-      style={{ backgroundColor: mflColors.card, borderColor: mflColors.border }}
-    >
-      <AppText className="text-sm font-semibold text-foreground" numberOfLines={1}>
-        {title}
-      </AppText>
-      <AppText className="mt-0.5 text-xs text-muted" numberOfLines={2}>
-        {subtitle}
-      </AppText>
-    </Pressable>
-  );
-}
